+++ mocks/ordersMapped.ts
@@ -1,0 +1,353 @@
+export const ordersMapped = {
+  series: [
+    [1536139985000, 6.3],
+    [1541016524000, 3.63],
+    [1545002327000, 17.72],
+    [1547932530000, 7.27],
+    [1549127305000, 7.27],
+  ],
+  seriesItems: [
+    {
+      id: '3442871-2',
+      created: 1536139985000,
+      resourceUrl: 'https://api.discogs.com/marketplace/orders/3442871-2',
+      total: 7,
+      fee: 0.7,
+      items: [
+        {
+          id: 677267257,
+          price: {
+            value: 2,
+            currency: 'EUR',
+          },
+          media_condition: 'Very Good Plus (VG+)',
+          sleeve_condition: 'Generic',
+          condition_comments: '',
+          item_location: '',
+          private_comments: '',
+          release: {
+            id: 822281,
+            resource_url: 'https://api.discogs.com/releases/822281',
+            description:
+              'Cassino & Laben* Feat. Marina Puglisi - Lasting Memories (12")',
+            thumbnail:
+              'https://i.discogs.com/qux7GkBDmaT7KL2J-hhrFbICz8URzFozI3Jm5K5paLk/rs:fit/g:sm/q:40/h:300/w:300/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9SLTgyMjI4/MS0xMjExMTI3NjU5/LmpwZWc.jpeg',
+          },
+        },
+        {
+          id: 677269800,
+          price: {
+            value: 1,
+            currency: 'EUR',
+          },
+          media_condition: 'Very Good (VG)',
+          sleeve_condition: 'Good (G)',
+          condition_comments: '',
+          item_location: '',
+          private_comments: '',
+          release: {
+            id: 405634,
+            resource_url: 'https://api.discogs.com/releases/405634',
+            description: 'Eelke Kleijn - Deeper Depths EP (12", EP)',
+            thumbnail:
+              'https://i.discogs.com/FHVsEWrqsXjNWxBLwHw8_jh_mveF_s6r1xImiIucuDs/rs:fit/g:sm/q:40/h:300/w:300/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9SLTQwNTYz/NC0xMTEzMjA5MTM5/LmpwZw.jpeg',
+          },
+        },
+        {
+          id: 677269823,
+          price: {
+            value: 3,
+            currency: 'EUR',
+          },
+          media_condition: 'Very Good (VG)',
+          sleeve_condition: 'Generic',
+          condition_comments: '',
+          item_location: '',
+          private_comments: '',
+          release: {
+            id: 471344,
+            resource_url: 'https://api.discogs.com/releases/471344',
+            description: 'Eelke Kleijn - Consequences / We Resist (12")',
+            thumbnail:
+              'https://i.discogs.com/u7-JJq6qP6gh1M6r0_1d5IdngsUo1XlkXTczKyEDTEQ/rs:fit/g:sm/q:40/h:300/w:300/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9SLTQ3MTM0/NC0xNTIyMDA2MDU4/LTU0MzQuanBlZw.jpeg',
+          },
+        },
+        {
+          id: 677270529,
+          price: {
+            value: 1,
+            currency: 'EUR',
+          },
+          media_condition: 'Very Good (VG)',
+          sleeve_condition: 'Very Good Plus (VG+)',
+          condition_comments: '',
+          item_location: '',
+          private_comments: '',
+          release: {
+            id: 442231,
+            resource_url: 'https://api.discogs.com/releases/442231',
+            description: 'Peter McCowan - EP (12")',
+            thumbnail:
+              'https://i.discogs.com/3_GqY_kPFq2NM1QUn3KSbaNPs2-uc6r6WOwFzKZQlJY/rs:fit/g:sm/q:40/h:300/w:300/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9SLTQ0MjIz/MS0xMzI4Mjc5MjMz/LmpwZWc.jpeg',
+          },
+        },
+      ],
+    },
+    {
+      id: '3442871-4',
+      created: 1541016524000,
+      resourceUrl: 'https://api.discogs.com/marketplace/orders/3442871-4',
+      total: 4,
+      fee: 0.37,
+      items: [
+        {
+          id: 677270574,
+          price: {
+            value: 4,
+            currency: 'EUR',
+          },
+          media_condition: 'Very Good Plus (VG+)',
+          sleeve_condition: 'Very Good (VG)',
+          condition_comments: '',
+          item_location: '',
+          private_comments: '',
+          release: {
+            id: 523940,
+            resource_url: 'https://api.discogs.com/releases/523940',
+            description: 'The Quasar* - October + Remixes (12")',
+            thumbnail:
+              'https://i.discogs.com/zFJdxdIOCT8v4xNdAKa_J_B4F4pkG11Hqnhm_TvNS64/rs:fit/g:sm/q:40/h:300/w:300/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9SLTUyMzk0/MC0xMTg5MDE0OTMy/LmpwZWc.jpeg',
+          },
+        },
+      ],
+    },
+    {
+      id: '3442871-5',
+      created: 1545002327000,
+      resourceUrl: 'https://api.discogs.com/marketplace/orders/3442871-5',
+      total: 19.5,
+      fee: 1.78,
+      items: [
+        {
+          id: 677268073,
+          price: {
+            value: 3,
+            currency: 'EUR',
+          },
+          media_condition: 'Very Good Plus (VG+)',
+          sleeve_condition: 'Generic',
+          condition_comments: '',
+          item_location: '',
+          private_comments: '',
+          release: {
+            id: 560488,
+            resource_url: 'https://api.discogs.com/releases/560488',
+            description: 'Dousk - Wrist Slap / Estrange (12")',
+            thumbnail:
+              'https://i.discogs.com/nGPFJeyMlp3h2Oad2UptKTse0yzgkzZ-9QhOCRqj6go/rs:fit/g:sm/q:40/h:300/w:300/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9SLTU2MDQ4/OC0xMzc1NDU5ODY5/LTQ1MjQuanBlZw.jpeg',
+          },
+        },
+        {
+          id: 677270418,
+          price: {
+            value: 1,
+            currency: 'EUR',
+          },
+          media_condition: 'Very Good Plus (VG+)',
+          sleeve_condition: 'Very Good Plus (VG+)',
+          condition_comments: '',
+          item_location: '',
+          private_comments: '',
+          release: {
+            id: 349900,
+            resource_url: 'https://api.discogs.com/releases/349900',
+            description: 'Orchid - Everlasting E.P. (12", EP)',
+            thumbnail:
+              'https://i.discogs.com/bcZebCiPF-UABI_S0KW7Yjyz4BDO5Z5tHcqb0S8jJLY/rs:fit/g:sm/q:40/h:300/w:300/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9SLTM0OTkw/MC0xMTA4NzU5NTkz/LmpwZw.jpeg',
+          },
+        },
+        {
+          id: 677270437,
+          price: {
+            value: 2,
+            currency: 'EUR',
+          },
+          media_condition: 'Very Good Plus (VG+)',
+          sleeve_condition: 'Very Good Plus (VG+)',
+          condition_comments: '',
+          item_location: '',
+          private_comments: '',
+          release: {
+            id: 464878,
+            resource_url: 'https://api.discogs.com/releases/464878',
+            description: 'Orchid - Technicolour EP (Part 1) (12")',
+            thumbnail:
+              'https://i.discogs.com/kfrsVvr43VhEGP3qI3Y_C_0qo6KIKWxxX8fjLwCqAs4/rs:fit/g:sm/q:40/h:300/w:300/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9SLTQ2NDg3/OC0xMTg5MzczODE4/LmpwZWc.jpeg',
+          },
+        },
+        {
+          id: 677270719,
+          price: {
+            value: 1.5,
+            currency: 'EUR',
+          },
+          media_condition: 'Very Good Plus (VG+)',
+          sleeve_condition: 'Generic',
+          condition_comments: '',
+          item_location: '',
+          private_comments: '',
+          release: {
+            id: 299663,
+            resource_url: 'https://api.discogs.com/releases/299663',
+            description:
+              'Subsky / Muzo* - Let It Flow / Particular People (12")',
+            thumbnail:
+              'https://i.discogs.com/b71_EoP2pH5bftgCPkpMpRSpvf0OKBaidjM3jGAPBx4/rs:fit/g:sm/q:40/h:300/w:300/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9SLTI5OTY2/My0xNjQ1NzE5NDU1/LTc5MjMuanBlZw.jpeg',
+          },
+        },
+        {
+          id: 677270993,
+          price: {
+            value: 3,
+            currency: 'EUR',
+          },
+          media_condition: 'Very Good Plus (VG+)',
+          sleeve_condition: 'Generic',
+          condition_comments: '',
+          item_location: '',
+          private_comments: '',
+          release: {
+            id: 749562,
+            resource_url: 'https://api.discogs.com/releases/749562',
+            description: 'V-Sag - Thespian (12")',
+            thumbnail:
+              'https://i.discogs.com/mxllCoQFinDXthqUwyD0Pg3JvxrPojiTSMaZTPLp4Xs/rs:fit/g:sm/q:40/h:300/w:300/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9SLTc0OTU2/Mi0xMTkwNDYwMzgw/LmpwZWc.jpeg',
+          },
+        },
+        {
+          id: 677271057,
+          price: {
+            value: 2.5,
+            currency: 'EUR',
+          },
+          media_condition: 'Very Good Plus (VG+)',
+          sleeve_condition: 'Very Good Plus (VG+)',
+          condition_comments: '',
+          item_location: '',
+          private_comments: '',
+          release: {
+            id: 700650,
+            resource_url: 'https://api.discogs.com/releases/700650',
+            description: 'V-Sag Meets dPen - Dark Inside (12")',
+            thumbnail:
+              'https://i.discogs.com/WA_QSebcaP-6BqecuDshS3KusMU_0DxAQGnRi3Npgsk/rs:fit/g:sm/q:40/h:300/w:300/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9SLTcwMDY1/MC0xNTU5OTI4Mjc2/LTMyMTUuanBlZw.jpeg',
+          },
+        },
+        {
+          id: 677271090,
+          price: {
+            value: 2.5,
+            currency: 'EUR',
+          },
+          media_condition: 'Very Good Plus (VG+)',
+          sleeve_condition: 'Generic',
+          condition_comments: '',
+          item_location: '',
+          private_comments: '',
+          release: {
+            id: 651168,
+            resource_url: 'https://api.discogs.com/releases/651168',
+            description:
+              'Various - Back To Back Sessions Vol.1 (Sampler) (Disc One) (12", Smplr)',
+            thumbnail:
+              'https://i.discogs.com/scG4zTNhfy8IwHRueSIp_6irNS6Xa3U_-9pZhsBGKs4/rs:fit/g:sm/q:40/h:300/w:300/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9SLTY1MTE2/OC0xMjM2MjY5OTcw/LmpwZWc.jpeg',
+          },
+        },
+        {
+          id: 843927416,
+          price: {
+            value: 4,
+            currency: 'EUR',
+          },
+          media_condition: 'Near Mint (NM or M-)',
+          sleeve_condition: 'Generic',
+          condition_comments: '',
+          item_location: '',
+          private_comments: '',
+          release: {
+            id: 520920,
+            resource_url: 'https://api.discogs.com/releases/520920',
+            description:
+              'Hawaii / Plastique Vision - Setting Up / Barely Away (12")',
+            thumbnail:
+              'https://i.discogs.com/zhreT-Zy2XexGiTpCUnQcUFl2U6tk0bQKesQqz_GCPI/rs:fit/g:sm/q:40/h:300/w:300/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9SLTUyMDky/MC0xMTg5MDE0MDY2/LmpwZWc.jpeg',
+          },
+        },
+      ],
+    },
+    {
+      id: '3442871-6',
+      created: 1547932530000,
+      resourceUrl: 'https://api.discogs.com/marketplace/orders/3442871-6',
+      total: 8,
+      fee: 0.73,
+      items: [
+        {
+          id: 854025605,
+          price: {
+            value: 8,
+            currency: 'EUR',
+          },
+          media_condition: 'Near Mint (NM or M-)',
+          sleeve_condition: 'Near Mint (NM or M-)',
+          condition_comments: '',
+          item_location: '',
+          private_comments: '',
+          release: {
+            id: 377926,
+            resource_url: 'https://api.discogs.com/releases/377926',
+            description: 'Ridgewalkers Feat. El - Find (12", Pic)',
+            thumbnail:
+              'https://i.discogs.com/gDXin9-g6M6BpzMl59cFJSr0_LlpdQ6Nt7NYTGr_fdY/rs:fit/g:sm/q:40/h:300/w:300/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9SLTM3Nzky/Ni0xNTUwMzM5Njc4/LTkzNjcuanBlZw.jpeg',
+          },
+        },
+      ],
+    },
+    {
+      id: '3442871-7',
+      created: 1549127305000,
+      resourceUrl: 'https://api.discogs.com/marketplace/orders/3442871-7',
+      total: 8,
+      fee: 0.73,
+      items: [
+        {
+          id: 854025439,
+          price: {
+            value: 8,
+            currency: 'EUR',
+          },
+          media_condition: 'Very Good (VG)',
+          sleeve_condition: 'Very Good Plus (VG+)',
+          condition_comments: '',
+          item_location: '',
+          private_comments: '',
+          release: {
+            id: 43284,
+            resource_url: 'https://api.discogs.com/releases/43284',
+            description:
+              'Madonna - What It Feels Like For A Girl (12", Single)',
+            thumbnail:
+              'https://i.discogs.com/xBPWXV8yKimCFqkLmfGZTr-NrhVtsW51aJLpjMXmpTA/rs:fit/g:sm/q:40/h:300/w:300/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9SLTQzMjg0/LTE1ODkxMTI0MzMt/MjU2OS5qcGVn.jpeg',
+          },
+        },
+      ],
+    },
+  ],
+  ordersTotal: 101,
+  ordersSent: 10,
+  ordersCancelled: 5,
+  ordersRefund: 1,
+  itemsSold: 100,
+  totalAmount: 100,
+  username: 'testuser',
+}