+++ types/OrdersPagination.ts
@@ -1,3 +1,10 @@
 export interface OrdersPagination {
+  page: number
   pages: number
+  per_page: number
+  items: number
+  urls: {
+    last: string
+    next: string
+  }
 }