+++ CHANGELOG.md
@@ -1,9 +1,16 @@
-
 # Change Log
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [2.0.1] - 2022-06-19
+
+### Added
+
+- unit tests for the frontend (more to be added)
+- small styling improvements
+- complete types for [OrdersPagination](./types/OrdersPagination.ts)
 
 ## [2.0.0] - 2022-06-18